Improve icon file structure
