+++ apps/spruce/src/pages/waterfall/icons/WhiteGear.ts
@@ -1,0 +1,1 @@
+export default `url('data:image/svg+xml,<svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" clip-rule="evenodd" d="M9.93081 1.93388C9.65253 1.81882 9.33165 1.90792 9.15253 2.14999L8.82493 2.59272C8.26818 2.50733 7.70987 2.51034 7.16817 2.59419L6.83969 2.15146C6.66027 1.90962 6.33927 1.82093 6.06114 1.93635L5.07578 2.34523C4.79765 2.46065 4.63375 2.75055 4.67827 3.04838L4.75976 3.59359C4.31784 3.91791 3.92142 4.31108 3.58869 4.76557L3.04389 4.68486C2.74601 4.64073 2.45632 4.80499 2.34126 5.08328L1.93363 6.06915C1.81857 6.34743 1.90767 6.66832 2.14974 6.84743L2.59247 7.17503C2.50709 7.73179 2.5101 8.29011 2.59395 8.8318L2.15123 9.16027C1.90939 9.3397 1.8207 9.66069 1.93611 9.93883L2.345 10.9242C2.46042 11.2023 2.75032 11.3662 3.04814 11.3217L3.59335 11.2402C3.91767 11.6821 4.31083 12.0785 4.76533 12.4113L4.68462 12.9561C4.64048 13.254 4.80475 13.5436 5.08303 13.6587L6.06891 14.0663C6.34719 14.1814 6.66807 14.0923 6.84719 13.8502L7.17479 13.4075C7.73155 13.4929 8.28986 13.4899 8.83156 13.406L9.16003 13.8487C9.33946 14.0906 9.66045 14.1793 9.93859 14.0638L10.9239 13.655C11.2021 13.5395 11.366 13.2496 11.3215 12.9518L11.24 12.4066C11.6819 12.0823 12.0783 11.6891 12.411 11.2346L12.9558 11.3154C13.2537 11.3595 13.5434 11.1952 13.6585 10.9169L14.0661 9.93106C14.1811 9.65277 14.092 9.33189 13.85 9.15278L13.4072 8.82518C13.4926 8.26842 13.4896 7.7101 13.4058 7.1684L13.8485 6.83994C14.0903 6.66051 14.179 6.33951 14.0636 6.06138L13.6547 5.07603C13.5393 4.79789 13.2494 4.634 12.9516 4.67851L12.4064 4.76C12.0821 4.31808 11.6889 3.92167 11.2344 3.58894L11.3151 3.04413C11.3592 2.74625 11.195 2.45656 10.9167 2.3415L9.93081 1.93388ZM10.2235 8.91951C9.71576 10.1476 8.30856 10.7316 7.08045 10.2238C5.85235 9.71601 5.26841 8.3088 5.77619 7.0807C6.28396 5.8526 7.69117 5.26866 8.91927 5.77643C10.1474 6.2842 10.7313 7.69141 10.2235 8.91951Z" fill="%23ffffff"/></svg>')`;
