+++ apps/spruce/src/pages/waterfall/icons/WhiteX.ts
@@ -1,0 +1,1 @@
+export default `url('data:image/svg+xml,<svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M11.8891 3.40379C11.4986 3.01327 10.8654 3.01327 10.4749 3.40379L8.00005 5.87867L5.52518 3.40379C5.13465 3.01327 4.50149 3.01327 4.11096 3.40379L3.40386 4.1109C3.01333 4.50143 3.01333 5.13459 3.40386 5.52511L5.87873 7.99999L3.40386 10.4749C3.01333 10.8654 3.01333 11.4986 3.40386 11.8891L4.11096 12.5962C4.50149 12.9867 5.13465 12.9867 5.52518 12.5962L8.00005 10.1213L10.4749 12.5962C10.8654 12.9867 11.4986 12.9867 11.8891 12.5962L12.5962 11.8891C12.9868 11.4986 12.9868 10.8654 12.5962 10.4749L10.1214 7.99999L12.5962 5.52511C12.9868 5.13459 12.9868 4.50143 12.5962 4.1109L11.8891 3.40379Z" fill="%23ffffff"/></svg>')`;
